refactor: rename object modules
