+++ io_scene_xray/object_/export.py
@@ -6,7 +6,7 @@ import mathutils
 
 from ..xray_io import ChunkedWriter, PackedWriter
 from ..xray_motions import export_motions, MATRIX_BONE_INVERTED
-from .fmt_object import Chunks
+from .format_ import Chunks
 from ..utils import is_exportable_bone, find_bone_exportable_parent, AppError, \
     convert_object_to_space_bmesh, calculate_mesh_bbox, gen_texture_name, is_helper_object
 from ..utils import BAD_VTX_GROUP_NAME
