+++ io_scene_xray/object_/import_.py
@@ -7,7 +7,7 @@ import bpy
 import mathutils
 
 from ..xray_io import ChunkedReader, PackedReader
-from .fmt_object import Chunks
+from .format_ import Chunks
 from ..plugin_prefs import PropObjectMeshSplitByMaterials
 from ..utils import BAD_VTX_GROUP_NAME, plugin_version_number, AppError
 from ..xray_motions import import_motions, MATRIX_BONE, MATRIX_BONE_INVERTED
